# release-23.2: colexecerror: avoid debug.Stack in CatchVectorizedRuntimeError