--- a/pkg/sql/sem/builtins/fixed_oids.go
+++ b/pkg/sql/sem/builtins/fixed_oids.go
@@ -2513,6 +2513,7 @@ var builtinOidsArray = []string{
 	2602: `crdb_internal.execute_internally(query: string, session_bound: bool, overrides: string) -> string`,
 	2603: `crdb_internal.execute_internally(query: string, overrides: string, use_session_txn: bool) -> string`,
 	2604: `crdb_internal.execute_internally(query: string, session_bound: bool, overrides: string, use_session_txn: bool) -> string`,
+	2608: `crdb_internal.force_panic(msg: string, mode: string) -> int`,
 }
 
 var builtinOidsBySignature map[string]oid.Oid